PR feedback: - let's restrict the scope of our route equivalency check to root override only - move our tests in with rest of url_helpers
