+++ actionview/lib/action_view/helpers/url_helper.rb
@@ -606,10 +606,12 @@ module ActionView
           request_uri = +"#{request.protocol}#{request.host_with_port}#{request_uri}"
         end
 
+        url_is_root = is_root_route_equivalent(url_string, request_uri)
+
         remove_trailing_slash!(url_string)
         remove_trailing_slash!(request_uri)
 
-        url_string == request_uri || compare_recognized(url_string, request_uri)
+        (url_string == request_uri) || url_is_root
       end
 
       if RUBY_VERSION.start_with?("2.7")
@@ -625,8 +627,8 @@ module ActionView
         end
       end
 
-      def compare_recognized(path1, path2)
-        _routes.recognize_path_spec(path1) == _routes.recognize_path_spec(path2)
+      def is_root_route_equivalent(path1, path2)
+        path1 == '/' && !(request.fullpath.index("?")) && _routes.recognize_path_spec(path1) == _routes.recognize_path_spec(path2)
       end
 
       # Creates an SMS anchor link tag to the specified +phone_number+. When the
