+++ actionview/test/template/url_helper_test.rb
@@ -1351,3 +1351,75 @@ class PolymorphicSessionsControllerTest < ActionController::TestCase
     assert_equal %{/workshops/1/sessions/1.json\n<a href="/workshops/1/sessions/1.json">Session</a>}, @response.body
   end
 end
+
+class CurrentPageRootTest < ActiveSupport::TestCase
+  attr_accessor :controller, :request
+
+  routes = ActionDispatch::Routing::RouteSet.new
+  routes.draw do
+    root "posts#index"
+    resources :posts, only: :index
+  end
+
+  include ActionView::Helpers::UrlHelper
+  include routes.url_helpers
+
+  def request_for_url(url, opts = {})
+    env = Rack::MockRequest.env_for("http://www.example.com#{url}", opts)
+    ActionDispatch::Request.new(env)
+  end
+
+  def test_returns_success
+    @request = request_for_url("/posts", method: :head)
+    assert current_page?(controller: "posts", action: "index")
+  end
+end
+
+class CurrentPageRootOnlyTest < ActiveSupport::TestCase
+  attr_accessor :controller, :request
+
+  routes = ActionDispatch::Routing::RouteSet.new
+  routes.draw do
+    resources :posts, only: :index
+  end
+
+  include ActionView::Helpers::UrlHelper
+  include routes.url_helpers
+
+  def request_for_url(url, opts = {})
+    env = Rack::MockRequest.env_for("http://www.example.com#{url}", opts)
+    ActionDispatch::Request.new(env)
+  end
+
+  def test_returns_success
+    @request = request_for_url("/posts", method: :head)
+    assert !current_page?(controller: "posts", action: "index")
+  end
+end
+
+class CurrentPageRootOnlyTest < ActiveSupport::TestCase
+  attr_accessor :controller, :request
+
+  routes = ActionDispatch::Routing::RouteSet.new
+  routes.draw do
+    resources :posts, only: :index
+  end
+
+  include ActionView::Helpers::UrlHelper
+  include routes.url_helpers
+
+  def request_for_url(url, opts = {})
+    env = Rack::MockRequest.env_for("http://www.example.com#{url}", opts)
+    ActionDispatch::Request.new(env)
+  end
+
+  def test_returns_success
+    @request = request_for_url("/posts", method: :head)
+    assert current_page?(controller: "posts", action: "index")
+  end
+
+  def test_root_returns_no_success
+    @request = request_for_url("/", method: :head)
+    assert !current_page?(controller: "posts", action: "index")
+  end
+end
